power/test_power_basics.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test for Power API</title>
</head>
<body>
<script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>

<script type="application/javascript">

power = navigator.mozPower;
ok(power, "Should be able to access power manager with permission.");

</script>
</pre>
</body>
</html>
